requirements.txt
# ── Orchestration ────────────────────────────────────────────────────────────
apache-airflow[amazon,postgres,apache.kafka]==2.8.1
apache-airflow-providers-amazon==8.10.0

# ── Data Processing ──────────────────────────────────────────────────────────
pyspark==3.5.0
pandas==2.1.4
pyarrow==14.0.2
numpy==1.26.3

# ── AWS ──────────────────────────────────────────────────────────────────────
boto3==1.34.14
awswrangler==3.6.0          # Pandas on AWS (S3, Redshift, Glue Catalog)
s3fs==2023.12.2

# ── Kafka ────────────────────────────────────────────────────────────────────
confluent-kafka==2.3.0
kafka-python==2.0.2

# ── dbt ──────────────────────────────────────────────────────────────────────
dbt-core==1.7.4
dbt-postgres==1.7.4         # Local dev with PostgreSQL
dbt-redshift==1.7.4         # Production with Redshift

# ── Data Quality ─────────────────────────────────────────────────────────────
great-expectations==0.18.12

# ── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
redshift-connector==2.0.918

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.0
faker==22.2.0
loguru==0.7.2
click==8.1.7
pydantic==2.5.3
tenacity==8.2.3             # Retry logic
croniter==2.0.1

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==7.4.4
pytest-cov==4.1.0
moto[s3,redshift]==4.2.14   # Mock AWS services
